+++ sdk/storage/Azure.Storage.DataMovement.Blobs/tests/BlobContainerClientExtensionsTests.cs
@@ -66,7 +66,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
 
             if (addTransferOptions)
             {
-                await client.StartUploadDirectoryAsync(directoryPath, new BlobContainerClientTransferOptions
+                await client.UploadDirectoryAsync(WaitUntil.Started, directoryPath, new BlobContainerClientTransferOptions
                 {
                     BlobContainerOptions = new() { BlobDirectoryPrefix = blobDirectoryPrefix },
                     TransferOptions = options
@@ -74,7 +74,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
             }
             else
             {
-                await client.StartUploadDirectoryAsync(directoryPath, blobDirectoryPrefix);
+                await client.UploadDirectoryAsync(WaitUntil.Started, directoryPath, blobDirectoryPrefix);
             }
 
             Assert.IsTrue(assertionComplete);
@@ -113,7 +113,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
 
             if (addTransferOptions)
             {
-                await client.StartDownloadToDirectoryAsync(directoryPath, new BlobContainerClientTransferOptions
+                await client.DownloadToDirectoryAsync(WaitUntil.Started, directoryPath, new BlobContainerClientTransferOptions
                 {
                     BlobContainerOptions = new() { BlobDirectoryPrefix = blobDirectoryPrefix },
                     TransferOptions = options
@@ -121,7 +121,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
             }
             else
             {
-                await client.StartDownloadToDirectoryAsync(directoryPath, blobDirectoryPrefix);
+                await client.DownloadToDirectoryAsync(WaitUntil.Started, directoryPath, blobDirectoryPrefix);
             }
 
             Assert.IsTrue(assertionComplete);